Support non-multilib setup
The configure script checks for a common header file
in the default gcc include path in order to detect
the support for i386.
The evenual result is saved in .config.in.
The linux target in the Makefile now depends on that
configuration file in order to enable or disable i386
support at compile time.

Signed-off-by: Olivier Tilmans <olivier.tilmans@uclouvain.be>